+++ test/unit/fuzz-tools.test.js
@@ -517,6 +517,7 @@ import * as islLymphedemaV539 from '../../lib/isl-lymphedema-v539.js';
 import * as ishltRejectionV540 from '../../lib/ishlt-rejection-v540.js';
 import * as rachs1V541 from '../../lib/rachs1-v541.js';
 import * as twstrsSeverityV542 from '../../lib/twstrs-severity-v542.js';
+import * as saveScoreV543 from '../../lib/save-score-v543.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -979,6 +980,7 @@ const MODULES = {
   'ishlt-rejection-v540.js': ishltRejectionV540,
   'rachs1-v541.js': rachs1V541,
   'twstrs-severity-v542.js': twstrsSeverityV542,
+  'save-score-v543.js': saveScoreV543,
   'tb-testing.js': tbTesting,
 };